pyproject.toml
[tool.poetry]
name = "django-python-service"
version = "0.1.0"
description = "Django Python Service"
authors = ["Ayush Pandey <ayushpandey.iitkgp@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.8"
django = "4.2.7"
psycopg2 = "^2.9.9"
djangorestframework = "3.14.0"
djangorestframework-api-key = "2.3.0"
drf-yasg = "1.21.6"
django-debug-toolbar = "4.2.0"
deepl = "^1.18.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.11.2"
pre-commit = "^3.8.0"
pytest = "8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
